Rename AuthContext, add more docs
